requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.3.0
billiard==3.3.0.23
blis==0.4.1
catalogue==1.0.0
celery==3.1.24
certifi==2020.6.20
chardet==3.0.4
cymem==2.0.4
dataclasses==0.7
Django==3.0.5
djangorestframework==3.12.1
djongo==1.3.3
en-core-web-sm==2.2.0
idna==2.10
importlib-metadata==2.0.0
kombu==3.0.37
murmurhash==1.0.4
numpy==1.19.4
plac==1.1.3
preshed==3.0.4
pymongo==3.11.0
pytz==2020.4
redis==2.10.6
requests==2.24.0
spacy==2.3.2
sqlparse==0.2.4
srsly==1.0.3
thinc==7.4.1
tqdm==4.51.0
urllib3==1.25.11
wasabi==0.8.0
zipp==3.4.0
